PROJECTFORGE-2366 fix NPEs and improve the precision of the calculation of the monthly salary

+++ projectforge-business/src/main/java/org/projectforge/business/fibu/api/EmployeeService.java
@@ -1,5 +1,6 @@
 package org.projectforge.business.fibu.api;
 
+import java.math.BigDecimal;
 import java.util.Calendar;
 
 import org.projectforge.business.fibu.EmployeeDO;
@@ -28,6 +29,6 @@ public interface EmployeeService extends IPersistenceService<EmployeeDO>, IDao<E
 
   boolean isEmployeeActive(EmployeeDO employee);
 
-  Float getMonthlySalary(EmployeeDO employee, Calendar selectedDate);
+  BigDecimal getMonthlySalary(EmployeeDO employee, Calendar selectedDate);
 
 }
